+++ stac_fastapi/sfeos_helpers/stac_fastapi/sfeos_helpers/mappings.py
@@ -295,7 +295,14 @@ _ES_INDEX_NAME_UNSUPPORTED_CHARS_TABLE = str.maketrans(
 ITEM_INDICES = f"{ITEMS_INDEX_PREFIX}*,-*kibana*,-{COLLECTIONS_INDEX}*"
 
 DEFAULT_SORT = {
-    "properties.datetime": {"order": "desc"},
+    "properties.datetime": {
+        "order": "desc",
+        "missing": 0,
+    },
+    "properties.start_datetime": {
+        "order": "desc",
+        "missing": 0,
+    },
     "id": {"order": "desc"},
     "collection": {"order": "desc"},
 }
@@ -303,7 +310,7 @@ DEFAULT_SORT = {
 ES_ITEMS_SETTINGS = {
     "index": {
         "sort.field": list(DEFAULT_SORT.keys()),
-        "sort.order": [v["order"] for v in DEFAULT_SORT.values()],
+        "sort.order": [v["order"] for v in DEFAULT_SORT.values()],  # type: ignore
         "mapping.coerce": COERCE_GLOBAL,
     }
 }